requirements.txt
fastapi
uvicorn[standard]
openai
httpx
python-multipart
beautifulsoup4
lxml
selenium
webdriver-manager
pydantic
python-dotenv
